# docs/api/openapi.yaml
info:
  description: 'Human-facing REST API for QuantDinger. Agent integrations use `/api/agent/v1`
    — see `docs/agent/agent-openapi.json`. Conventions: `docs/architecture/API_CONVENTIONS.md`.'
  contact:
    name: QuantDinger
    url: https://github.com/OpenByteInc/quantdinger
  license:
    name: See repository LICENSE
  title: QuantDinger Web API
  version: 1.0.0
components:
  securitySchemes:
    HumanJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'User session JWT from `POST /api/auth/login`. Send as `Authorization:
        Bearer <token>`.'
    AgentToken:
      type: http
      scheme: bearer
      bearerFormat: qd_agent_xxx
      description: Agent Gateway token for `/api/agent/v1/*` only. See docs/agent/AGENT_QUICKSTART.md.
  schemas:
...
    HumanSuccessEnvelope:
      type: object
      properties:
        code:
          type: integer
          description: 1 on success
          example: 1
        msg:
          type: string
          example: success
        data:
          nullable: true
      additionalProperties: false
    HumanErrorEnvelope:
      type: object
      properties:
        code:
          type: integer
          description: 0 on error
          example: 0
        msg:
          type: string
        data:
          nullable: true
      additionalProperties: false
...
servers:
- url: http://localhost:5000
  description: Backend direct (python run.py)
- url: http://localhost:8888
  description: Docker Compose (via reverse proxy)
paths:
...
  /api/account/positions:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: L1 account position mirror for Quick Trade / asset views.
      tags:
      - Account
      operationId: getApiAccountPositions
      x-visibility: internal
...
  /api/alpaca/accounts:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      tags:
      - Alpaca
      operationId: getApiAlpacaAccounts
      summary: Get Api Alpaca Accounts
      x-visibility: internal
  /api/alpaca/status:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Get Alpaca connection status.
      tags:
      - Alpaca
      operationId: getApiAlpacaStatus
      x-visibility: internal
  /api/alpaca/connect:
    post:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Connect to Alpaca.
      description: 'Request body:

        apiKey (required): API key (PK prefix = paper, AK = live)

        secretKey (required): Secret key

        paper (optional, default true): Use paper trading'
      tags:
      - Alpaca
      operationId: postApiAlpacaConnect
      x-visibility: internal
  /api/alpaca/disconnect:
    post:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Disconnect from Alpaca.
      tags:
      - Alpaca
      operationId: postApiAlpacaDisconnect
      x-visibility: internal
  /api/alpaca/account:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Get Alpaca account information.
      tags:
      - Alpaca
      operationId: getApiAlpacaAccount
      x-visibility: internal
  /api/alpaca/positions:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Get Alpaca open positions.
      tags:
      - Alpaca
      operationId: getApiAlpacaPositions
      x-visibility: internal
  /api/alpaca/orders:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Get Alpaca recent orders or only open orders.
      tags:
      - Alpaca
      operationId: getApiAlpacaOrders
      x-visibility: internal
  /api/alpaca/order:
    post:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Place an Alpaca order.
      description: 'Request body:

        symbol (required): Ticker, e.g. AAPL

        side (required): buy or sell

        quantity or notional (required): Share quantity or USD amount; mutually exclusive

        marketType (optional): USStock or crypto (default USStock)

        orderType (optional): market or limit (default market)

        price (required for limit): Limit price

        extendedHours (optional): Allow extended-hours limit orders'
      tags:
      - Alpaca
      operationId: postApiAlpacaOrder
      x-visibility: internal
  /api/alpaca/order/{order_id}:
    delete:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Cancel an Alpaca order by ID.
      tags:
      - Alpaca
      operationId: deleteApiAlpacaOrderOrderId
      parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: integer
      x-visibility: internal
    parameters:
    - in: path
      name: order_id
      required: true
      schema:
        type: string
        minLength: 1
  /api/alpaca/quote/{symbol}:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Get a real-time Alpaca quote.
      description: 'Query params:

        marketType (optional): USStock or crypto (default USStock)'
      tags:
      - Alpaca
      operationId: getApiAlpacaQuoteSymbol
      parameters:
      - name: symbol
        in: path
        required: true
        schema:
          type: string
      x-visibility: internal
    parameters:
    - in: path
      name: symbol
      required: true
      schema:
        type: string
        minLength: 1
...
openapi: 3.0.3